Paramatise the little finger origin.
May not be the right approach, as may need to fix up the axis of the the LFJ5
joing as well. Which with this setup will need the axis block passing in as
well. However all we do is flip the x cord on the origin and axis so a reflect
param (1|-1) might be better as we will never need more than one little finger.
Can test that after creating a mirrored mesh for the joint.